Learning about event delegation
I was going to put an event handler for each individual category for the nav bar but realised thats a really inefficient way to code. So, I googled and found out about event delegation.

My main problem is having the img rotate unique to the target event and not have other images rotate.