SUCESSS!!! set interval madness
I figured out how to continuously add and remove class of the image gallery using setInterval for addClass and within, use setTimeout for removeClass. I also figured out that I needed a local reference for i because the value of i can change as it loops through the image gallery which can lead to unexpected results.