Finishing up next button code
I got the next button code to work but the problem is when I click the img inside of the next button, the button1 or button2 code does not activate. I tried to use querySelector on the img id but that still does not activate the code. Bug after bugggg