Almost done with mini button animation
I am almost done with the animation with the mini buttons. I used indexOf and arrays for the mini buttons and used that to also update index which is also shared with sliderImage so it updates as the mini buttons are updated. Aka I can do classList.add 'outline' to both at the same time in real time.

The rest is trying to see why the class 'outline' is removed when not clicking the mini buttons and little bug with the slider animation when clicking the mini buttons to animate rather than the next buttons.