Problem with transition of game-gallery-items
I finished up with having the transition effect of when you hover over an game-gallery-item, the opacity of items other is reduced. I used classList.add within an addEventListener of mouseover and mouseout.

The problem is with the transition where when you click a button-choice, the transition is supposed to make a game-gallery-item appear one at a time but I have been having some problems with it. This is due to the opacity uses of the mouseover and mouseout events combined with the need to use opacity for the transition effect I am trying to create